+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.0] - 2025-12-05
+
+### Breaking Changes
+
+- Renamed `valTownAdapter` to `sqliteAdapter` - the adapter works with any
+  SQLite driver using the `execute({ sql, args })` pattern (Val.Town, libSQL,
+  Turso, etc.)
+
+### Migration
+
+Replace:
+
+```typescript
+import { valTownAdapter } from "@tijs/atproto-storage";
+```
+
+With:
+
+```typescript
+import { sqliteAdapter } from "@tijs/atproto-storage";
+```
+
 ## [0.1.1] - 2025-11-27
 
 ### Fixed